Issue #SC-1026 feat: Druid Query Processor refactoring
